feat(dsg): renaming createGraphQLFieldDecorator
