+++ packages/amplication-data-service-generator/src/server/resource/dto/graphql-field-decorator/index.ts
@@ -1,0 +1,1 @@
+export * from "./create-graphql-field-decorator";
